pneumonia-api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
tensorflow==2.15.0
pillow==10.0.1
python-multipart==0.0.6
numpy==1.24.3
pydantic==2.4.2
gunicorn==21.2.0
